Fold message construction into the build module
assembly.py held one function with one call site and no test file of its own; building the
TestBatch messages is the last step of build_test_batches. The shared jobs() test builder moves
to tests.helpers.batching so both test modules can use it.
